fix behavior if selected theme is not in SPOTLIGHT_THEMES
- move theme methods into separate helper
- push themed view method up from starlight repo
